VideoContainer/Sender/requirements.txt
picamera2~=0.1.0
opencv-python~=4.12.0.88
numpy~=2.2.1
python-dotenv>=1.0.1